# api-docs/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/dict/ajuster": {
            "get": {
                "summary": "Get a list of adjusters",
                "operationId": "aaecaccd6ab4e4ab982bd8ac3c03d828",
                "responses": {
                    "200": {
                        "description": "List of adjusteres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "body": {
                                                "type": "string"
                                            },
                                            "in_group": {
                                                "type": "string"
                                            },
                                            "branch_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dict/adjuster/create": {
            "post": {
                "summary": "Create a new adjuster",
                "operationId": "e27670ba62d58198e07c6065b4339ce3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "Body of the adjuster",
                                        "type": "string"
                                    },
                                    "in_group": {
                                        "description": "Group of the adjuster",
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "description": "Adjuster is a member of Branch",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "in_group": {
                                            "type": "string"
                                        },
                                        "branch_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dict/adjuster/{id}": {
            "get": {
                "summary": "Get adjuster details by ID",
                "operationId": "23093bb5aaaaa98464e0c9d621df1858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the adjuster",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "in_group": {
                                            "type": "string"
                                        },
                                        "branch_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Adjuster not found"
                    }
                }
            },
            "put": {
                "summary": "Update adjuster by ID",
                "operationId": "6b223c1423eb36d8e47fab9b8ed1e3ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the adjuster to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "New title for the adjuster",
                                        "type": "string"
                                    },
                                    "in_group": {
                                        "description": "New goup for the adjuster",
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "description": "New branch_id for the adjuster",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "in_group": {
                                            "type": "string"
                                        },
                                        "branch_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete adjuster by ID",
                "operationId": "90dc2255310e91a1b819b210e35f2a64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the adjuster to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
        "/dict/branch": {
            "get": {
                "summary": "Get a list of branches",
                "operationId": "ea55d7d7db15d59d8d1f6b6c3d2d9263",
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "body": {
                                                "type": "string"
                                            },
                                            "prefix": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dict/branch/create": {
            "post": {
                "summary": "Create a new branch",
                "operationId": "660626169431501338ccc43e30e83038",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "Body of the branch",
                                        "type": "string"
                                    },
                                    "prefix": {
                                        "description": "Prefix to name for branch members",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dict/branch/{id}": {
            "get": {
                "summary": "Get branch details by ID",
                "operationId": "0120f0b0ae6367b9913e9641aec7486a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the branch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "prefix": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "put": {
                "summary": "Update branch by ID",
                "operationId": "9b3608eadfa0266c2d6b7f65799af8bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the branch to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "New title for the branch",
                                        "type": "string"
                                    },
                                    "prefix": {
                                        "description": "New prefix for the branch members",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "prefix": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete branch by ID",
                "operationId": "bbf5ebe9b9e11be720032da3394b8f68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the branch to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        }
    }
}